Updated Commission Calculator
This was my very first project, and it's been a few weeks of learning since I wrote this code. So as a little challenge to myself, I have decided to re-write this code to make it more efficient, and to require user input. 

This is the first chunk of code - this will ask the user how many days they worked this week and how many sales they made each day - then create a list to store this information.

The next step is to repeat this method to create an equal length array with the number of lead generations per day.